Tech Talks Framework (#219)
This pull request implements a framework for Awesome Learning to host recordings of Tech Talks. Right now there are few talks on React Boston; it will be easy to add many more of these in bulk once the framework is there, but I wanted to get just a few to start.